Gracefully handle existing package with unclear version
This commit implements a check that can trigger especially for
VCS-packages where the version is previously unknown due to it being
constructed from VCS-information such as commit id. As the version is
previously unknown, blinky has no way of checking if that package is
already present in the package cache, so the package is built anyways.
Afterwards, blinky can check this (and does now).

+++ package_tree.py
@@ -259,7 +259,11 @@ class Package:
 		if fullpkgnames:
 			for fpn in fullpkgnames:
 				self.built_pkgs.append(fpn)
-				shutil.move(os.path.join(self.srcpkg.srcdir, fpn), self.ctx.cachedir)
+				if os.path.exists(os.path.join(self.ctx.cachedir, fpn)):
+					if not os.path.isfile(os.path.join(self.ctx.cachedir, fpn)):
+						utils.logerr(None, 0, "Something (that's not a file) is shadowing package {} in cache directory {}".format((fpn), self.cachedir))
+				else:
+					shutil.move(os.path.join(self.srcpkg.srcdir, fpn), self.ctx.cachedir)
 		else:
 			utils.logerr(None, "No package file found in builddir for {}, aborting this subtree".format(self.name))
 			return False
